programming - refactored sampling to use SamplingContext for (1) thread-local Random instance and (2) current instantiation of sampledVars without need to frequently reallocate

+++ src/BNA/src/bna/bnlib/sampling/SamplingController.java
@@ -8,15 +8,15 @@ package bna.bnlib.sampling;
  * To control the sampling process, possibly in context of more threads.
  */
 public class SamplingController {
-    private int maxSamples;
+    private long maxSamples;
     private boolean stopFlag;
     
-    public SamplingController(int maxSamples) {
+    public SamplingController(long maxSamples) {
         this.maxSamples = maxSamples;
         this.stopFlag = false;
     }
     
-    public int maxSamples() {
+    public long maxSamples() {
         return this.maxSamples;
     }
